+++ packages/dara-core/js/shared/index.tsx
@@ -40,5 +40,6 @@ export {
     resolveValue,
     useVariableValue,
 } from './interactivity';
+export { useEventBus, EventBus, EventCapturer } from './event-bus/event-bus';
 export { default as Wrapper } from './wrapper/wrapper';
 export type { RawCssProp };