+++ packages/dara-core/tests/js/dynamic-component.spec.tsx
@@ -4,7 +4,9 @@ import { rest } from 'msw';
 import * as React from 'react';
 
 import { FallbackCtx } from '@/shared/context';
+import { EventCapturer } from '@/shared/event-bus/event-bus';
 import { clearRegistries_TEST } from '@/shared/interactivity/store';
+import { DaraEventMap } from '@/types/event-types';
 
 import { DynamicComponent, useAction, useVariable } from '../../js/shared';
 import { Action, DerivedVariable, SingleVariable, Variable } from '../../js/types';
@@ -98,6 +100,55 @@ describe('DynamicComponent', () => {
                     'TestComponent2: {"uid":"uid","values":{"data":{"test":{"__ref":"Variable:uid"}},"lookup":{"Variable:uid":"test"}},"ws_channel":"uid"}'
                 )
             ).toBeInstanceOf(HTMLDivElement);
+        });
+    });
+
+    it('should publish EventBus notification on server component render', async () => {
+        const receivedData: Array<DaraEventMap['SERVER_COMPONENT_LOADED']> = [];
+
+        const { getByText } = wrappedRender(
+            <EventCapturer
+                onEvent={(event) => {
+                    if (event.type === 'SERVER_COMPONENT_LOADED') {
+                        receivedData.push(event.data);
+                    }
+                }}
+            >
+                <DynamicComponent
+                    component={{
+                        name: 'TestComponent2',
+                        props: {
+                            // Check that dynamic kwargs get passed to the backend
+                            dynamic_kwargs: {
+                                test: {
+                                    __typename: 'Variable',
+                                    default: 'test',
+                                    uid: 'uid',
+                                },
+                            },
+                        },
+                        uid: 'uid',
+                    }}
+                />
+            </EventCapturer>
+        );
+        await waitFor(() => {
+            expect(getByText(/TestComponent2/)).toBeInstanceOf(HTMLDivElement);
+        });
+
+        const value = getByText(/TestComponent2/).textContent;
+
+        // single event bus notification should come through
+        expect(receivedData).toHaveLength(1);
+        expect(receivedData[0]).toEqual({
+            uid: 'uid',
+            name: 'TestComponent2',
+            value: {
+                name: 'RawString',
+                props: {
+                    content: value,
+                },
+            }
         });
     });